realized "customers who bought this book also bought these" mechanism

+++ src/main/java/bookshop/dao/book/CommentDao.java
@@ -6,9 +6,9 @@ import bookshop.model.user.User;
 import java.util.List;
 
 public interface CommentDao {
-	List<Comment> getAll(Long bookId);
+	List<Comment> getAll(Integer bookId);
 	Comment getCommentById(Long id);
-	void persistComment(Long bookId, Comment comment);
+	void persistComment(Integer bookId, Comment comment);
 	void deleteCommentById(Long id);
 	void updateComment(Long id);
 	void likeComment(Long id, User user);
