+++ src/main/java/bookshop/service/book/CommentService.java
@@ -6,9 +6,9 @@ import bookshop.model.book.Comment;
 import bookshop.model.user.User;
 
 public interface CommentService {
-	List<Comment> getAll(Long bookId);
+	List<Comment> getAll(Integer bookId);
 	Comment getCommentById(Long id);
-	void persistComment(Long bookId, Comment comment);
+	void persistComment(Integer bookId, Comment comment);
 	void deleteCommentById(Long id);
 	void updateComment(Long id);
 	void likeComment(Long id, User user);
