+++ src/main/java/bookshop/service/user/UserService.java
@@ -20,9 +20,9 @@ public interface UserService {
      
     boolean isUserSSOUnique(Integer id, String sso);
 
-	void addBookToBasket(Long bookId, String ssoId);
+	void addBookToBasket(Integer bookId, String ssoId);
 
-	void removeBookFromBasket(Long bookId, String ssoId);
+	void removeBookFromBasket(Integer bookId, String ssoId);
 
 	void commitPurchase(String ssoId);
